+++ codenames/pyproject.toml
@@ -14,6 +14,7 @@ graphene-federation = "0.2.0"
 openai = "^1.2.0"
 python-dotenv = "^1.0.0"
 flask-cors = "^4.0.0"
+upstash-redis = "^0.15.0"
 
 
 [build-system]
