+++ codenames/requirements.txt
@@ -1,14 +1,19 @@
+aiohttp==3.9.0 ; python_version >= "3.11" and python_version < "4.0"
+aiosignal==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
 aniso8601==7.0.0 ; python_version >= "3.11" and python_version < "4.0"
 annotated-types==0.6.0 ; python_version >= "3.11" and python_version < "4.0"
 anyio==3.7.1 ; python_version >= "3.11" and python_version < "4.0"
+attrs==23.1.0 ; python_version >= "3.11" and python_version < "4.0"
 blinker==1.7.0 ; python_version >= "3.11" and python_version < "4.0"
 certifi==2023.7.22 ; python_version >= "3.11" and python_version < "4.0"
+charset-normalizer==3.3.2 ; python_version >= "3.11" and python_version < "4.0"
 click==8.1.7 ; python_version >= "3.11" and python_version < "4.0"
 colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
 distro==1.8.0 ; python_version >= "3.11" and python_version < "4.0"
 flask-cors==4.0.0 ; python_version >= "3.11" and python_version < "4.0"
 flask-graphql==2.0.1 ; python_version >= "3.11" and python_version < "4.0"
 flask==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
+frozenlist==1.4.0 ; python_version >= "3.11" and python_version < "4.0"
 graphene-federation==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
 graphene==2.1.9 ; python_version >= "3.11" and python_version < "4.0"
 graphql-core==2.3.2 ; python_version >= "3.11" and python_version < "4.0"
@@ -21,14 +26,19 @@ idna==3.4 ; python_version >= "3.11" and python_version < "4.0"
 itsdangerous==2.1.2 ; python_version >= "3.11" and python_version < "4.0"
 jinja2==3.1.2 ; python_version >= "3.11" and python_version < "4.0"
 markupsafe==2.1.3 ; python_version >= "3.11" and python_version < "4.0"
+multidict==6.0.4 ; python_version >= "3.11" and python_version < "4.0"
 openai==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
 promise==2.3 ; python_version >= "3.11" and python_version < "4.0"
 pydantic-core==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
 pydantic==2.4.2 ; python_version >= "3.11" and python_version < "4.0"
 python-dotenv==1.0.0 ; python_version >= "3.11" and python_version < "4.0"
+requests==2.31.0 ; python_version >= "3.11" and python_version < "4.0"
 rx==1.6.3 ; python_version >= "3.11" and python_version < "4.0"
 six==1.16.0 ; python_version >= "3.11" and python_version < "4.0"
 sniffio==1.3.0 ; python_version >= "3.11" and python_version < "4.0"
 tqdm==4.66.1 ; python_version >= "3.11" and python_version < "4.0"
 typing-extensions==4.8.0 ; python_version >= "3.11" and python_version < "4.0"
+upstash-redis==0.15.0 ; python_version >= "3.11" and python_version < "4.0"
+urllib3==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
 werkzeug==3.0.1 ; python_version >= "3.11" and python_version < "4.0"
+yarl==1.9.3 ; python_version >= "3.11" and python_version < "4.0"
