+++ frontend/package.json
@@ -20,6 +20,7 @@
     "graphql": "^16.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.20.0",
     "serve": "^14.2.1"
   },
   "devDependencies": {
