+++ frontend/src/gql/gql.ts
@@ -13,10 +13,9 @@ import { TypedDocumentNode as DocumentNode } from '@graphql-typed-document-node/
  * Therefore it is highly recommended to use the babel or swc plugin for production.
  */
 const documents = {
-    "\n  query GetBoardQueryDocument {\n    game {\n      board {\n        wordValue\n        typeValue\n        isRevealed\n        position {\n          x\n          y\n        }\n      }\n      turn\n      turnCount\n      winner\n    }\n  }\n": types.GetBoardQueryDocumentDocument,
-    "\n  query GetCurrentClueQueryDocument {\n    game {\n      currentClue {\n        word\n        number\n      }\n    }\n  }\n": types.GetCurrentClueQueryDocumentDocument,
-    "\n  query GetGameRecapQueryDocument {\n    game {\n      redClues {\n        word\n        number\n        reasoning\n      }\n      blueClues {\n        word\n        number\n        reasoning\n      }\n    }\n  }\n": types.GetGameRecapQueryDocumentDocument,
-    "\n  mutation InitializeGameMutationDocument {\n    initializeGame {\n      game {\n        board {\n          wordValue\n          typeValue\n          isRevealed\n          position {\n            x\n            y\n          }\n        }\n        turn\n        turnCount\n        winner\n      }\n    }\n  }\n": types.InitializeGameMutationDocumentDocument,
+    "\n  query GetBoardQueryDocument($roomId: ID!) {\n    game(roomId: $roomId) {\n      board {\n        wordValue\n        typeValue\n        isRevealed\n        position {\n          x\n          y\n        }\n      }\n      turn\n      turnCount\n      winner\n    }\n  }\n": types.GetBoardQueryDocumentDocument,
+    "\n  query GetCurrentClueQueryDocument($roomId: ID!) {\n    game(roomId: $roomId) {\n      currentClue {\n        word\n        number\n      }\n    }\n  }\n": types.GetCurrentClueQueryDocumentDocument,
+    "\n  query GetGameRecapQueryDocument($roomId: ID!) {\n    game(roomId: $roomId) {\n      redClues {\n        word\n        number\n        reasoning\n      }\n      blueClues {\n        word\n        number\n        reasoning\n      }\n    }\n  }\n": types.GetGameRecapQueryDocumentDocument,
     "\n  mutation GuessCardMutationDocument($position: PositionInput!) {\n    guessCard(position: $position) {\n      ok\n    }\n  }\n": types.GuessCardMutationDocumentDocument,
     "\n  mutation EndTurnMutationDocument {\n    endTurn {\n      ok\n    }\n  }\n": types.EndTurnMutationDocumentDocument,
     "\n  mutation EndGameMutationDocument {\n    endGame {\n      ok\n    }\n  }\n": types.EndGameMutationDocumentDocument,
@@ -40,19 +39,15 @@ export function graphql(source: string): unknown;
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query GetBoardQueryDocument {\n    game {\n      board {\n        wordValue\n        typeValue\n        isRevealed\n        position {\n          x\n          y\n        }\n      }\n      turn\n      turnCount\n      winner\n    }\n  }\n"): (typeof documents)["\n  query GetBoardQueryDocument {\n    game {\n      board {\n        wordValue\n        typeValue\n        isRevealed\n        position {\n          x\n          y\n        }\n      }\n      turn\n      turnCount\n      winner\n    }\n  }\n"];
+export function graphql(source: "\n  query GetBoardQueryDocument($roomId: ID!) {\n    game(roomId: $roomId) {\n      board {\n        wordValue\n        typeValue\n        isRevealed\n        position {\n          x\n          y\n        }\n      }\n      turn\n      turnCount\n      winner\n    }\n  }\n"): (typeof documents)["\n  query GetBoardQueryDocument($roomId: ID!) {\n    game(roomId: $roomId) {\n      board {\n        wordValue\n        typeValue\n        isRevealed\n        position {\n          x\n          y\n        }\n      }\n      turn\n      turnCount\n      winner\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query GetCurrentClueQueryDocument {\n    game {\n      currentClue {\n        word\n        number\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetCurrentClueQueryDocument {\n    game {\n      currentClue {\n        word\n        number\n      }\n    }\n  }\n"];
+export function graphql(source: "\n  query GetCurrentClueQueryDocument($roomId: ID!) {\n    game(roomId: $roomId) {\n      currentClue {\n        word\n        number\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetCurrentClueQueryDocument($roomId: ID!) {\n    game(roomId: $roomId) {\n      currentClue {\n        word\n        number\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
-export function graphql(source: "\n  query GetGameRecapQueryDocument {\n    game {\n      redClues {\n        word\n        number\n        reasoning\n      }\n      blueClues {\n        word\n        number\n        reasoning\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetGameRecapQueryDocument {\n    game {\n      redClues {\n        word\n        number\n        reasoning\n      }\n      blueClues {\n        word\n        number\n        reasoning\n      }\n    }\n  }\n"];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n  mutation InitializeGameMutationDocument {\n    initializeGame {\n      game {\n        board {\n          wordValue\n          typeValue\n          isRevealed\n          position {\n            x\n            y\n          }\n        }\n        turn\n        turnCount\n        winner\n      }\n    }\n  }\n"): (typeof documents)["\n  mutation InitializeGameMutationDocument {\n    initializeGame {\n      game {\n        board {\n          wordValue\n          typeValue\n          isRevealed\n          position {\n            x\n            y\n          }\n        }\n        turn\n        turnCount\n        winner\n      }\n    }\n  }\n"];
+export function graphql(source: "\n  query GetGameRecapQueryDocument($roomId: ID!) {\n    game(roomId: $roomId) {\n      redClues {\n        word\n        number\n        reasoning\n      }\n      blueClues {\n        word\n        number\n        reasoning\n      }\n    }\n  }\n"): (typeof documents)["\n  query GetGameRecapQueryDocument($roomId: ID!) {\n    game(roomId: $roomId) {\n      redClues {\n        word\n        number\n        reasoning\n      }\n      blueClues {\n        word\n        number\n        reasoning\n      }\n    }\n  }\n"];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
